+++ Mile.Helpers/Mile.Helpers.CppBase.cpp
@@ -9,3 +9,39 @@
  */
 
 #include "Mile.Helpers.CppBase.h"
+
+#include <cstdarg>
+
+std::wstring Mile::VFormatWideString(
+    _In_z_ _Printf_format_string_ wchar_t const* const Format,
+    _In_z_ _Printf_format_string_ va_list ArgList)
+{
+    // Check the argument list.
+    if (Format)
+    {
+        // Get the length of the format result.
+        std::size_t nLength =
+            static_cast<std::size_t>(::_vscwprintf(Format, ArgList)) + 1;
+
+        // Allocate for the format result.
+        std::wstring Buffer(nLength + 1, L'\0');
+
+        // Format the string.
+        int nWritten = ::_vsnwprintf_s(
+            &Buffer[0],
+            Buffer.size(),
+            nLength,
+            Format,
+            ArgList);
+
+        if (nWritten > 0)
+        {
+            // If succeed, resize to fit and return result.
+            Buffer.resize(nWritten);
+            return Buffer;
+        }
+    }
+
+    // If failed, return an empty string.
+    return std::wstring();
+}