+++ Mile.Helpers/Mile.Helpers.CppBase.h
@@ -13,8 +13,20 @@
 #ifndef MILE_WINDOWS_HELPERS_CPPBASE
 #define MILE_WINDOWS_HELPERS_CPPBASE
 
+#include <string>
+
 namespace Mile
 {
+    /**
+     * @brief Write formatted data to a wide characters string, assumed UTF-16
+     *        in Windows.
+     * @param Format Format-control string.
+     * @param ArgList Pointer to list of optional arguments to be formatted.
+     * @return A formatted string if successful, an empty string otherwise.
+    */
+    std::wstring VFormatWideString(
+        _In_z_ _Printf_format_string_ wchar_t const* const Format,
+        _In_z_ _Printf_format_string_ va_list ArgList);
 
 }