+++ ReleaseNotes.md
@@ -31,3 +31,4 @@
 - Add MileIsCurrentProcessElevated function.
 - Add MileAllowNonClientDefaultDrawingForWindow function.
 - Add Mile::WinRT::GetLocalizedString method.
+- Add Mile::VFormatWideString method.